flail around: Broke it again, no idea how.

+++ BtDamageResolver/src/Actors/Logic/Interfaces/ILogicUnit.cs
@@ -137,7 +137,7 @@ public interface ILogicUnit
     /// <param name="targetNumber">The target number.</param>
     /// <param name="weaponEntry">The weapon entry to calculate projection for.</param>
     /// <returns>A tuple containing the estimated and maximum ammo usage.</returns>
-    Task<(double Estimate, int Max)> ProjectAmmo(int targetNumber, WeaponEntry weaponEntry);
+    Task<(decimal Estimate, int Max)> ProjectAmmo(int targetNumber, WeaponEntry weaponEntry);
 
     /// <summary>
     /// Calculates projected heat generation for weapon.
@@ -145,7 +145,7 @@ public interface ILogicUnit
     /// <param name="targetNumber">The target number.</param>
     /// <param name="weaponEntry">The weapon entry to calculate projection for.</param>
     /// <returns>A tuple containing the estimated and maximum heat generation.</returns>
-    Task<(double Estimate, int Max)> ProjectHeat(int targetNumber, WeaponEntry weaponEntry);
+    Task<(decimal Estimate, int Max)> ProjectHeat(int targetNumber, WeaponEntry weaponEntry);
 
     /// <summary>
     /// Transforms a cluster roll based on unit type and possible other properties.
